feat: dfdaemon is divided into download and upload
