+++ src/grpc/dfdaemon_download.rs
@@ -15,25 +15,23 @@
  */
 
 use crate::shutdown;
-use crate::storage;
 use crate::task;
 use crate::utils::http::hashmap_to_headermap;
 use crate::Result as ClientResult;
-use dragonfly_api::common::v2::{Piece, Task};
+use dragonfly_api::common::v2::Task;
 use dragonfly_api::dfdaemon::v2::{
-    dfdaemon_client::DfdaemonClient as DfdaemonGRPCClient,
-    dfdaemon_server::{Dfdaemon, DfdaemonServer as DfdaemonGRPCServer},
-    DeleteTaskRequest, DownloadPieceRequest, DownloadPieceResponse, DownloadTaskRequest,
-    DownloadTaskResponse, StatTaskRequest as DfdaemonStatTaskRequest, SyncPiecesRequest,
-    SyncPiecesResponse, UploadTaskRequest,
+    dfdaemon_download_client::DfdaemonDownloadClient as DfdaemonDownloadGRPCClient,
+    dfdaemon_download_server::{
+        DfdaemonDownload, DfdaemonDownloadServer as DfdaemonDownloadGRPCServer,
+    },
+    DeleteTaskRequest, DownloadTaskRequest, DownloadTaskResponse,
+    StatTaskRequest as DfdaemonStatTaskRequest, UploadTaskRequest,
 };
 use dragonfly_api::scheduler::v2::StatTaskRequest as SchedulerStatTaskRequest;
 use std::borrow::BorrowMut;
-use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
-use tokio::io::AsyncReadExt;
 use tokio::net::{UnixListener, UnixStream};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::{ReceiverStream, UnixListenerStream};
@@ -46,91 +44,13 @@ use tonic_health::server::HealthReporter;
 use tower::service_fn;
 use tracing::{error, info, instrument, Instrument, Span};
 
-// DfdaemonUploadServer is the grpc server of the upload.
-pub struct DfdaemonUploadServer {
-    // addr is the address of the grpc server.
-    addr: SocketAddr,
-
-    // service is the grpc service of the dfdaemon.
-    service: DfdaemonGRPCServer<DfdaemonServerHandler>,
-
-    // health_reporter is the health reporter of the grpc server.
-    health_reporter: HealthReporter,
-
-    // shutdown is used to shutdown the grpc server.
-    shutdown: shutdown::Shutdown,
-
-    // _shutdown_complete is used to notify the grpc server is shutdown.
-    _shutdown_complete: mpsc::UnboundedSender<()>,
-}
-
-// DfdaemonUploadServer implements the grpc server of the upload.
-impl DfdaemonUploadServer {
-    // new creates a new DfdaemonUploadServer.
-    pub fn new(
-        addr: SocketAddr,
-        task: Arc<task::Task>,
-        health_reporter: HealthReporter,
-        shutdown: shutdown::Shutdown,
-        shutdown_complete_tx: mpsc::UnboundedSender<()>,
-    ) -> Self {
-        // Initialize the grpc service.
-        let service = DfdaemonGRPCServer::new(DfdaemonServerHandler { task })
-            .send_compressed(CompressionEncoding::Gzip)
-            .accept_compressed(CompressionEncoding::Gzip)
-            .max_decoding_message_size(usize::MAX);
-
-        Self {
-            addr,
-            service,
-            health_reporter,
-            shutdown,
-            _shutdown_complete: shutdown_complete_tx,
-        }
-    }
-
-    // run starts the upload server.
-    #[instrument(skip_all)]
-    pub async fn run(&mut self) {
-        // Register the reflection service.
-        let reflection = tonic_reflection::server::Builder::configure()
-            .register_encoded_file_descriptor_set(dragonfly_api::FILE_DESCRIPTOR_SET)
-            .build()
-            .unwrap();
-
-        // Clone the shutdown channel.
-        let mut shutdown = self.shutdown.clone();
-
-        // Borrow the health reporter.
-        let health_reporter = self.health_reporter.borrow_mut();
-
-        // Set the serving status of the upload grpc server.
-        health_reporter
-            .set_serving::<DfdaemonGRPCServer<DfdaemonServerHandler>>()
-            .await;
-
-        // Start upload grpc server.
-        info!("upload server listening on {}", self.addr);
-        Server::builder()
-            .add_service(reflection.clone())
-            .add_service(self.service.clone())
-            .serve_with_shutdown(self.addr, async move {
-                // Upload grpc server shutting down with signals.
-                let _ = shutdown.recv().await;
-                info!("upload grpc server shutting down");
-            })
-            .await
-            .unwrap();
-    }
-}
-
 // DfdaemonDownloadServer is the grpc unix server of the download.
 pub struct DfdaemonDownloadServer {
     // socket_path is the path of the unix domain socket.
     socket_path: PathBuf,
 
     // service is the grpc service of the dfdaemon.
-    service: DfdaemonGRPCServer<DfdaemonServerHandler>,
+    service: DfdaemonDownloadGRPCServer<DfdaemonDownloadServerHandler>,
 
     // health_reporter is the health reporter of the grpc server.
     health_reporter: HealthReporter,
@@ -153,10 +73,11 @@ impl DfdaemonDownloadServer {
         shutdown_complete_tx: mpsc::UnboundedSender<()>,
     ) -> Self {
         // Initialize the grpc service.
-        let service = DfdaemonGRPCServer::new(DfdaemonServerHandler { task: task.clone() })
-            .send_compressed(CompressionEncoding::Gzip)
-            .accept_compressed(CompressionEncoding::Gzip)
-            .max_decoding_message_size(usize::MAX);
+        let service =
+            DfdaemonDownloadGRPCServer::new(DfdaemonDownloadServerHandler { task: task.clone() })
+                .send_compressed(CompressionEncoding::Gzip)
+                .accept_compressed(CompressionEncoding::Gzip)
+                .max_decoding_message_size(usize::MAX);
 
         Self {
             socket_path,
@@ -184,7 +105,7 @@ impl DfdaemonDownloadServer {
 
         // Set the serving status of the download grpc server.
         health_reporter
-            .set_serving::<DfdaemonGRPCServer<DfdaemonServerHandler>>()
+            .set_serving::<DfdaemonDownloadGRPCServer<DfdaemonDownloadServerHandler>>()
             .await;
 
         // Start download grpc server with unix domain socket.
@@ -211,186 +132,15 @@ impl DfdaemonDownloadServer {
     }
 }
 
-// DfdaemonServerHandler is the handler of the dfdaemon grpc service.
-pub struct DfdaemonServerHandler {
+// DfdaemonDownloadServerHandler is the handler of the dfdaemon download grpc service.
+pub struct DfdaemonDownloadServerHandler {
     // task is the task manager.
     task: Arc<task::Task>,
 }
 
-// DfdaemonServerHandler implements the dfdaemon grpc service.
+// DfdaemonDownloadServerHandler implements the dfdaemon download grpc service.
 #[tonic::async_trait]
-impl Dfdaemon for DfdaemonServerHandler {
-    // SyncPiecesStream is the stream of the sync pieces response.
-    type SyncPiecesStream = ReceiverStream<Result<SyncPiecesResponse, Status>>;
-
-    // get_piece_numbers gets the piece numbers.
-    #[instrument(skip_all, fields(task_id))]
-    async fn sync_pieces(
-        &self,
-        request: Request<SyncPiecesRequest>,
-    ) -> Result<Response<Self::SyncPiecesStream>, Status> {
-        // Clone the request.
-        let request = request.into_inner();
-
-        // Get the task id from tae request.
-        let task_id = request.task_id;
-
-        // Span record the task id.
-        Span::current().record("task_id", task_id.clone());
-
-        // Get the interested piece numbers from the request.
-        let mut interested_piece_numbers = request.interested_piece_numbers.clone();
-
-        // Clone the task.
-        let task = self.task.clone();
-
-        // Initialize stream channel.
-        let (out_stream_tx, out_stream_rx) = mpsc::channel(128);
-        tokio::spawn(
-            async move {
-                loop {
-                    let mut has_started_piece = false;
-                    let mut finished_piece_numbers = Vec::new();
-                    for interested_piece_number in interested_piece_numbers.iter() {
-                        let piece = match task.piece.get(task_id.as_str(), *interested_piece_number)
-                        {
-                            Ok(Some(piece)) => piece,
-                            Ok(None) => continue,
-                            Err(err) => {
-                                error!("get piece metadata: {}", err);
-                                out_stream_tx
-                                    .send(Err(Status::internal(err.to_string())))
-                                    .await
-                                    .unwrap_or_else(|err| {
-                                        error!("send piece metadata to stream: {}", err);
-                                    });
-
-                                drop(out_stream_tx);
-                                return;
-                            }
-                        };
-
-                        // Send the piece metadata to the stream.
-                        if piece.is_finished() {
-                            info!("piece {} is finished", piece.number);
-                            out_stream_tx
-                                .send(Ok(SyncPiecesResponse {
-                                    piece_number: piece.number,
-                                }))
-                                .await
-                                .unwrap_or_else(|err| {
-                                    error!("send finished pieces to stream: {}", err);
-                                });
-
-                            // Add the finished piece number to the finished piece numbers.
-                            finished_piece_numbers.push(piece.number);
-                            continue;
-                        }
-
-                        // Check whether the piece is started.
-                        if piece.is_started() {
-                            has_started_piece = true;
-                        }
-                    }
-
-                    // Remove the finished piece numbers from the interested piece numbers.
-                    interested_piece_numbers
-                        .retain(|number| !finished_piece_numbers.contains(number));
-
-                    // If all the interested pieces are finished, return.
-                    if interested_piece_numbers.is_empty() {
-                        info!("all the interested pieces are finished");
-                        drop(out_stream_tx);
-                        return;
-                    }
-
-                    // If there is no started piece, return.
-                    if !has_started_piece {
-                        info!("there is no started piece");
-                        drop(out_stream_tx);
-                        return;
-                    }
-
-                    // Wait for the piece to be finished.
-                    tokio::time::sleep(storage::DEFAULT_WAIT_FOR_PIECE_FINISHED_INTERVAL).await;
-                }
-            }
-            .in_current_span(),
-        );
-
-        Ok(Response::new(ReceiverStream::new(out_stream_rx)))
-    }
-
-    // sync_pieces syncs the pieces.
-    #[instrument(skip_all, fields(task_id, piece_number))]
-    async fn download_piece(
-        &self,
-        request: Request<DownloadPieceRequest>,
-    ) -> Result<Response<DownloadPieceResponse>, Status> {
-        // Clone the request.
-        let request = request.into_inner();
-
-        // Clone the task.
-        let task = self.task.clone();
-
-        // Get the task id from the request.
-        let task_id = request.task_id;
-
-        // Span record the task id.
-        Span::current().record("task_id", task_id.as_str());
-
-        // Get the interested piece number from the request.
-        let piece_number = request.piece_number;
-
-        // Span record the piece number.
-        Span::current().record("piece_number", piece_number);
-
-        // Get the piece metadata from the local storage.
-        let piece = task
-            .piece
-            .get(task_id.as_str(), piece_number)
-            .map_err(|err| {
-                error!("get piece metadata from local storage: {}", err);
-                Status::internal(err.to_string())
-            })?
-            .ok_or_else(|| {
-                error!("piece metadata not found");
-                Status::not_found("piece metadata not found")
-            })?;
-
-        // Get the piece content from the local storage.
-        let mut reader = task
-            .piece
-            .download_from_local_peer_into_async_read(task_id.as_str(), piece_number)
-            .await
-            .map_err(|err| {
-                error!("read piece content from local storage: {}", err);
-                Status::internal(err.to_string())
-            })?;
-
-        // Read the content of the piece.
-        let mut content = Vec::new();
-        reader.read_to_end(&mut content).await.map_err(|err| {
-            error!("read piece content: {}", err);
-            Status::internal(err.to_string())
-        })?;
-
-        // Return the piece.
-        Ok(Response::new(DownloadPieceResponse {
-            piece: Some(Piece {
-                number: piece.number,
-                parent_id: piece.parent_id,
-                offset: piece.offset,
-                length: piece.length,
-                digest: piece.digest,
-                content: Some(content),
-                traffic_type: None,
-                cost: None,
-                created_at: None,
-            }),
-        }))
-    }
-
+impl DfdaemonDownload for DfdaemonDownloadServerHandler {
     // DownloadTaskStream is the stream of the download task response.
     type DownloadTaskStream = ReceiverStream<Result<DownloadTaskResponse, Status>>;
 
@@ -574,29 +324,16 @@ impl Dfdaemon for DfdaemonServerHandler {
     }
 }
 
-// DfdaemonClient is a wrapper of DfdaemonGRPCClient.
+// DfdaemonDownloadClient is a wrapper of DfdaemonDownloadGRPCClient.
 #[derive(Clone)]
-pub struct DfdaemonClient {
+pub struct DfdaemonDownloadClient {
     // client is the grpc client of the dfdaemon.
-    pub client: DfdaemonGRPCClient<Channel>,
+    pub client: DfdaemonDownloadGRPCClient<Channel>,
 }
 
-// DfdaemonClient implements the grpc client of the dfdaemon.
-impl DfdaemonClient {
-    // new creates a new DfdaemonClient.
-    pub async fn new(addr: String) -> ClientResult<Self> {
-        let channel = Channel::from_static(Box::leak(addr.into_boxed_str()))
-            .connect_timeout(super::CONNECT_TIMEOUT)
-            .connect()
-            .await?;
-        let client = DfdaemonGRPCClient::new(channel)
-            .send_compressed(CompressionEncoding::Gzip)
-            .accept_compressed(CompressionEncoding::Gzip)
-            .max_decoding_message_size(usize::MAX);
-        Ok(Self { client })
-    }
-
-    // new_unix creates a new DfdaemonClient with unix domain socket.
+// DfdaemonDownloadClient implements the grpc client of the dfdaemon download.
+impl DfdaemonDownloadClient {
+    // new_unix creates a new DfdaemonDownloadClient with unix domain socket.
     pub async fn new_unix(socket_path: PathBuf) -> ClientResult<Self> {
         // Ignore the uri because it is not used.
         let channel = Endpoint::try_from("http://[::]:50051")
@@ -605,36 +342,11 @@ impl DfdaemonClient {
                 UnixStream::connect(socket_path.clone())
             }))
             .await?;
-        let client = DfdaemonGRPCClient::new(channel)
+        let client = DfdaemonDownloadGRPCClient::new(channel)
             .send_compressed(CompressionEncoding::Gzip)
             .accept_compressed(CompressionEncoding::Gzip)
             .max_decoding_message_size(usize::MAX);
         Ok(Self { client })
-    }
-
-    // get_piece_numbers gets the piece numbers.
-    #[instrument(skip_all)]
-    pub async fn sync_pieces(
-        &self,
-        request: SyncPiecesRequest,
-    ) -> ClientResult<tonic::Response<tonic::codec::Streaming<SyncPiecesResponse>>> {
-        let request = Self::make_request(request);
-        let response = self.client.clone().sync_pieces(request).await?;
-        Ok(response)
-    }
-
-    // sync_pieces syncs the pieces.
-    #[instrument(skip_all)]
-    pub async fn download_piece(
-        &self,
-        request: DownloadPieceRequest,
-        timeout: Duration,
-    ) -> ClientResult<DownloadPieceResponse> {
-        let mut request = tonic::Request::new(request);
-        request.set_timeout(timeout);
-
-        let response = self.client.clone().download_piece(request).await?;
-        Ok(response.into_inner())
     }
 
     // download_task tells the dfdaemon to download the task.
