+++ src/grpc/health.rs
@@ -18,6 +18,7 @@ use crate::shutdown;
 use crate::{Error, Result};
 use std::net::SocketAddr;
 use tokio::sync::mpsc;
+use tokio::task::JoinSet;
 use tonic::transport::{Channel, Server};
 use tonic_health::pb::{
     health_client::HealthClient as HealthGRPCClient,
@@ -106,8 +107,53 @@ impl HealthClient {
     }
 
     // check checks the health of the server.
-    pub async fn check(&self, request: HealthCheckRequest) -> Result<HealthCheckResponse> {
-        let request = Self::make_request(request);
+    pub async fn check(&self) -> Result<HealthCheckResponse> {
+        let services = vec![
+            "dfdaemon.v2.DfdaemonDownload".to_string(),
+            "dfdaemon.v2.DfdaemonUpload".to_string(),
+        ];
+
+        let mut join_set = JoinSet::new();
+        for service in services {
+            let client = self.clone();
+            async fn check_service(
+                client: HealthClient,
+                service: String,
+            ) -> Result<HealthCheckResponse> {
+                client.check_service(service).await
+            }
+
+            join_set.spawn(check_service(client, service));
+        }
+
+        // Wait for all tasks to finish.
+        while let Some(message) = join_set.join_next().await {
+            match message {
+                Ok(check) => info!("check: {:?}", check),
+                Err(err) => return Err(err.into()),
+            };
+        }
+
+        Ok(HealthCheckResponse {
+            status: tonic_health::ServingStatus::Serving as i32,
+        })
+    }
+
+    // check_dfdaemon_download checks the health of the dfdaemon download service.
+    pub async fn check_dfdaemon_download(&self) -> Result<HealthCheckResponse> {
+        self.check_service("dfdaemon.v2.DfdaemonDownload".to_string())
+            .await
+    }
+
+    // check_dfdaemon_upload checks the health of the dfdaemon upload service.
+    pub async fn check_dfdaemon_upload(&self) -> Result<HealthCheckResponse> {
+        self.check_service("dfdaemon.v2.DfdaemonUpload".to_string())
+            .await
+    }
+
+    // check_service checks the health of the service.
+    async fn check_service(&self, service: String) -> Result<HealthCheckResponse> {
+        let request = Self::make_request(HealthCheckRequest { service });
         let response = self.client.clone().check(request).await?;
         Ok(response.into_inner())
     }
@@ -117,65 +163,5 @@ impl HealthClient {
         let mut request = tonic::Request::new(request);
         request.set_timeout(super::REQUEST_TIMEOUT);
         request
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use std::{net::SocketAddr, time::Duration};
-
-    use tokio::task::JoinHandle;
-    use tonic::transport::Server;
-    use tonic_health::{pb::HealthCheckRequest, ServingStatus};
-
-    use crate::grpc::health::HealthClient;
-
-    struct MockHealthServer {
-        addr: String,
-        handle: JoinHandle<()>,
-    }
-
-    impl Drop for MockHealthServer {
-        fn drop(&mut self) {
-            self.handle.abort()
-        }
-    }
-
-    async fn spawn_mock_app() -> MockHealthServer {
-        let (mut health_reporter, health_service) = tonic_health::server::health_reporter();
-        health_reporter
-            .set_service_status("test", ServingStatus::Serving)
-            .await;
-
-        let addr = SocketAddr::from(([127, 0, 0, 1], 8081));
-        let handle = tokio::spawn(async move {
-            Server::builder()
-                .add_service(health_service)
-                .serve(addr)
-                .await
-                .expect("failed to start server");
-        });
-        tokio::time::sleep(Duration::from_secs(1)).await;
-        return MockHealthServer {
-            handle,
-            addr: addr.to_string(),
-        };
-    }
-
-    #[tokio::test]
-    async fn test_get_avaliable_scheduler() {
-        let app = spawn_mock_app().await;
-        let health_client = HealthClient::new(&format!("http://{}", app.addr))
-            .await
-            .expect("failed to create health client");
-        let Ok(res) = health_client
-            .check(HealthCheckRequest {
-                service: String::new(),
-            })
-            .await
-        else {
-            panic!("failed to check health");
-        };
-        assert_eq!(res.status, ServingStatus::Serving as i32);
     }
 }
